+++ src/App.tsx
@@ -1,48 +1,18 @@
 
-
 import './App.css'
-
-import Highlight from '@tiptap/extension-highlight'
-import Typography from '@tiptap/extension-typography'
-import { EditorContent, useEditor } from '@tiptap/react'
-import StarterKit from '@tiptap/starter-kit'
 
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Editor from "./pages/editor";
+import Home from "./pages/Home";
 
 import React from 'react'
 
 export default () => {
-  const editor = useEditor({
-    extensions: [
-      StarterKit,
-      Highlight,
-      Typography,
-    ],
-    content: `
-    <p>
-      Markdown shortcuts make it easy to format the text while typing.
-    </p>
-    <p>
-      To test that, start a new line and type <code>#</code> followed by a space to get a heading. Try <code>#</code>, <code>##</code>, <code>###</code>, <code>####</code>, <code>#####</code>, <code>######</code> for different levels.
-    </p>
-    <p>
-      Those conventions are called input rules in Tiptap. Some of them are enabled by default. Try <code>></code> for blockquotes, <code>*</code>, <code>-</code> or <code>+</code> for bullet lists, or <code>\`foobar\`</code> to highlight code, <code>~~tildes~~</code> to strike text, or <code>==equal signs==</code> to highlight text.
-    </p>
-    <p>
-      You can overwrite existing input rules or add your own to nodes, marks and extensions.
-    </p>
-    <p>
-      For example, we added the <code>Typography</code> extension here. Try typing <code>(c)</code> to see how it’s converted to a proper © character. You can also try <code>-></code>, <code>>></code>, <code>1/2</code>, <code>!=</code>, or <code>--</code>.
-    </p>
-    `,
-  })
-
   return (
     <>
-      <EditorContent editor={editor} />
       <Router>
         <Routes>
+          <Route path="/" element={<Home />} />
           <Route path="/editor" element={<Editor />} />
         </Routes>
       </Router>
